requirements.txt
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.2
Pillow==9.5.0
pytorch_lightning==2.0.2
PyYAML==6.0
scikit-image==0.20.0
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.66.1
wandb==0.15.3
